Integrate @normenmueller's last commits in master
* Add @phaller's travis.yml

* Add ensime-related files to .gitignore

* Perform renaming of impl by backend
