+++ .gitignore
@@ -10,6 +10,8 @@ target/
 project/project/
 .idea/
 out/
+.ensime
+.ensime_cache/
 
 # wiki
 wiki
