+++ .travis.yml
@@ -1,8 +1,40 @@
+# Use Docker-based container (instead of OpenVZ)
+sudo: false
+
+cache:
+  directories:
+    - $HOME/.ivy2/cache
+
+    # At the moment, sbt 0.13.5 is preinstalled in Travis VM image,
+    # which fortunately corresponds to current scalaz settings.
+    # The line below can be used to cache a given sbt version.
+    # - $HOME/.sbt/launchers/0.13.x
+
+    # The line below is used to cache the scala version used by the build
+    # job, as these versions might be replaced after a Travis CI build
+    # environment upgrade (e.g. scala 2.11.2 could be replaced by scala 2.11.4).
+    - $HOME/.sbt/boot/scala-$TRAVIS_SCALA_VERSION
+
 language: scala
+
 scala:
   - 2.11.7
 
 jdk:
+  - openjdk7
+  - oraclejdk7
   - oraclejdk8
 
-script: sbt ++$TRAVIS_SCALA_VERSION scalastyle test
+notifications:
+  email:
+    - heather.miller@epfl.ch
+    - philipp.haller@a3.epfl.ch
+    - normen.mueller@gmail.com
+    - jorgevc@fastmail.es
+
+script:
+  - sbt ++$TRAVIS_SCALA_VERSION clean test
+
+  # Tricks to avoid unnecessary cache updates
+  - find $HOME/.sbt -name "*.lock" | xargs rm
+  - find $HOME/.ivy2 -name "ivydata-*.properties" | xargs rm
