+++ core/src/main/scala/Silo.scala
@@ -2,12 +2,6 @@ package fp
 
 import scala.concurrent.Future
 
-trait SiloFactory[T] {
-
-  def data: T
-
-}
-
 class Silo[T](private[fp] val data: T) {
 
   //def internalApply[A, V, B <: Traversable[V]](fun: A => B): LocalSilo[V, B] = {
