+++ core/src/main/scala/SiloFactory.scala
@@ -1,0 +1,7 @@
+package fp
+
+trait SiloFactory[T] {
+
+  def data: T
+
+}
