+++ core/src/main/scala/SiloRef.scala
@@ -2,34 +2,9 @@ package fp
 
 import com.typesafe.scalalogging.{ StrictLogging => Logging }
 import fp.core._
-import model.Populate
+import fp.model.{ Traverse, Traversed }
 
-import scala.concurrent.Future
-import scala.pickling._
-
-/** Interface implemented by [[SiloSystem]], the only two places
-  * from which you can get new silos.
-  */
-trait SiloRefFactory {
-
-  /** Upload a silo to `host` with the initialization process defined by `fun`.
-    *
-    * @tparam T Type of data to be populated
-    * @param fun Silo initialization logic
-    * @param at Target host of the the to be created silo
-    * @return [[fp.SiloRef]], which identifies the created silo
-    */
-  def populate[T](at: Host)(fun: () => Silo[T])(implicit pickler: Pickler[Populate[T]]): Future[SiloRef[T]]
-
-  /** Uploads a silo to a `Host` populated by `factory`.
-    *
-    * @param factory Silo initialization logic
-    * @param at Target host where the `Silo` will be created
-    * @return [[fp.SiloRef]], which identifies the created silo
-    */
-  final def populate[T](at: Host, factory: SiloFactory[T])(implicit pickler: Pickler[Populate[T]]): Future[SiloRef[T]] =
-    populate(at) { () => new Silo[T](factory.data) }
-}
+import scala.concurrent.{ ExecutionContext, Future }
 
 /** A `Silo` is uniquely identified by an `id`. */
 private[fp] case class SiloRefId(id: RefId, at: Host)
@@ -62,8 +37,7 @@ abstract class SiloRefAdapter[T] extends SiloRef[T] with Logging {
 
   import logger._
   import fp.core._
-
-  import Defaults._
+  import scala.pickling.Defaults._
   import scala.concurrent.ExecutionContext.Implicits.global
 
   protected def system: fp.Internals
@@ -71,14 +45,11 @@ abstract class SiloRefAdapter[T] extends SiloRef[T] with Logging {
   protected def node: Node
 
   override def send(): Future[T] = {
-    /*
     debug(s"Sending graph to host `${id.at}`...")
     system.request(id.at) { msgId => Traverse(msgId, node) } map {
       case Traversed(_, v) => v.asInstanceOf[T]
       case _ => throw new Exception(s"Computation at `${id.at}` failed.")
     }
-    */
-    ???
   }
 
 }
