+++ core/src/main/scala/backend/Server.scala
@@ -1,8 +1,8 @@
 package fp
-package impl
+package backend
 
 /** The internal API of a F-P server. */
-private[impl] trait Server extends Runnable {
+private[backend] trait Server extends Runnable {
 
   /** Host this server is running at */
   def at: Host
