+++ core/src/main/scala/backend/netty/Ask.scala
@@ -1,5 +1,5 @@
 package fp
-package impl
+package backend
 package netty
 
 import _root_.io.netty.channel.Channel
