+++ core/src/main/scala/backend/netty/Tell.scala
@@ -1,5 +1,5 @@
 package fp
-package impl
+package backend
 package netty
 
 import fp.model.Message
