+++ core/src/main/scala/backend/netty/codec.scala
@@ -1,5 +1,5 @@
 package fp
-package impl
+package backend
 package netty
 
 import java.io.ByteArrayOutputStream
