+++ core/src/main/scala/backend/netty/package.scala
@@ -1,5 +1,5 @@
 package fp
-package impl
+package backend
 
 import java.util.concurrent.CancellationException
