+++ samples/src/multi-jvm/scala/cs/Clients.scala
@@ -1,29 +1,27 @@
 package multijvm
 package cs
 
-import scala.concurrent.{ Await, ExecutionContext, Future, Promise }
-import ExecutionContext.Implicits.global
-import scala.concurrent.duration._
-import scala.util.{ Success, Failure }
-
-import com.typesafe.scalalogging.{ StrictLogging => Logging }
-
+import com.typesafe.scalalogging.{StrictLogging => Logging}
 import fp._
 
+import scala.concurrent.Await
+import scala.concurrent.duration._
+
 /** A silo system running in client mode can be understood as ''master'', or ''driver''.
-  * 
-  * All these terms have their raison d'être, i.e., all these terms, in general, state the fact that this node is for
-  * defining the control flow, the to be executed computations, and for sending those computations to respective silo
+  *
+  * All these terms have their raison d'être, i.e., all these terms, in general,
+  * state the fact that this node is for defining the control flow, the to be
+  * executed computations, and for sending those computations to respective silo
   * systems running in server mode.
-  */ 
-object ExampleMultiJvmClient extends AnyRef with Logging {
+  */
+object Clients extends AnyRef with Logging {
 
   private val summary = """
 In this talk, I'll present some of our ongoing work on a new programming model for asynchronous and distributed
 programming. For now, we call it "function-passing" or "function-passing style", and it can be thought of as an
 inversion of the actor model - keep your data stationary, send and apply your functionality (functions/spores) to that
 stationary data, and get typed communication all for free, all in a friendly collections/futures-like package!
-"""
+                        """
 
   private lazy val words: Array[String] = summary.replace('\n', ' ').split(" ")
 
@@ -34,7 +32,8 @@ stationary data, and get typed communication all for free, all in a friendly col
 
   val numLines = 10
 
-  // To be "siloed" data: Each string is a concatenation of 10 random words, separated by space
+  // To be "siloed" data: Each string is a concatenation of 10 random words,
+  // separated by space
 
   // Alternative 1: Define data via `() => Silo[T]`
   //val data = () => {
@@ -64,7 +63,7 @@ stationary data, and get typed communication all for free, all in a friendly col
 
   import logger._
 
-  def main(args: Array[String]): Unit = try { 
+  def main(args: Array[String]): Unit = try {
     /* Start a silo system in client mode. */
     val system = Await.result(SiloSystem(), 10.seconds)
     info(s"Silo system `${system.name}` up and running.")
@@ -74,7 +73,8 @@ stationary data, and get typed communication all for free, all in a friendly col
 
     /* Populate initial silo, i.e., upload initial data. */
     import scala.pickling.Defaults._
-    val ref: SiloRef[List[String]] = Await.result(system.populate(at, data), 10.seconds)    
+    val ref: SiloRef[List[String]] =
+      Await.result(system.populate(at, data), 10.seconds)
     info(s"Data populated at ${ref.id.at}")
 
     /* Force execution. */
@@ -86,10 +86,9 @@ stationary data, and get typed communication all for free, all in a friendly col
     //result foreach (info(_))
 
     Await.result(system.terminate, Duration.Inf)
-  } catch { case err: Throwable => 
+  } catch { case err: Throwable =>
     logger.error(s"Silo system terminated with error `${err.getMessage}`.")
-  } 
- 
+  }
+
 }
 
-// vim: set tw=120 ft=scala:
